[refactor] Image 도메인 패키지 생성 및 도메인 패키지별 파일 위치 리팩토링
