+++ yechef/src/main/java/sejong/capston/yechef/domain/Image/repository/ImageRepository.java
@@ -1,7 +1,7 @@
-package sejong.capston.yechef.domain.Recipe.repository;
+package sejong.capston.yechef.domain.Image.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sejong.capston.yechef.domain.Recipe.Image;
+import sejong.capston.yechef.domain.Image.Image;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
 }
