+++ yechef/src/main/java/sejong/capston/yechef/domain/Ingredient/repository/IngredientRepository.java
@@ -1,8 +1,8 @@
-package sejong.capston.yechef.domain.Recipe.repository;
+package sejong.capston.yechef.domain.Ingredient.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import sejong.capston.yechef.domain.Ingredients.Ingredient;
+import sejong.capston.yechef.domain.Ingredient.Ingredient;
 import sejong.capston.yechef.domain.Recipe.Recipe;
 
 public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
