+++ yechef/src/main/java/sejong/capston/yechef/domain/MemberRecipe/repository/MemberRecipeRepository.java
@@ -1,7 +1,8 @@
-package sejong.capston.yechef.domain.MemberRecipe;
+package sejong.capston.yechef.domain.MemberRecipe.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import sejong.capston.yechef.domain.MemberRecipe.MemberRecipe;
 
 public interface MemberRecipeRepository extends JpaRepository<MemberRecipe, Long> {
