+++ yechef/src/main/java/sejong/capston/yechef/global/s3/service/S3UploadService.java
@@ -7,9 +7,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import sejong.capston.yechef.domain.Recipe.Image;
+import sejong.capston.yechef.domain.Image.Image;
 import sejong.capston.yechef.domain.Recipe.Recipe;
-import sejong.capston.yechef.domain.Recipe.repository.ImageRepository;
+import sejong.capston.yechef.domain.Image.repository.ImageRepository;
 import sejong.capston.yechef.global.exception.BaseException;
 import sejong.capston.yechef.global.exception.error.ErrorCode;
 import software.amazon.awssdk.core.sync.RequestBody;
